Split BrowserLogs pipe launcher partials
Move platform-specific native launch logic into Windows and Unix partial classes while keeping the shared launcher entry points in the common file.
